requirements.txt
# Telegram Bot
python-telegram-bot==20.7
python-dotenv==1.0.0

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.23

# Data Analysis & Visualization
pandas==2.1.4
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.18.0

# AI/ML
google-generativeai==0.3.2

# Utilities
python-dateutil==2.8.2
pillow==10.1.0
